[chore] 단위 int -> float로 변경

+++ Assets/02.Scripts/Util/NumFormatter.cs
@@ -1,6 +1,6 @@
 public static class NumberFormatter
 {
-    public static string FormatIntNum(int number)
+    public static string FormatIntNum(float number)
     {
         if (number < 1000)
         {
